Backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pyjwt
flask-bcrypt
unidecode